Decompose amux-notify lib.rs into focused modules (fixes #87)
Splits the 1,127-line lib.rs into three modules:
- types.rs (85 LoC): AgentState, NotificationSource, FlashReason,
  WorkspaceStatus, Notification, PaneNotifyState
- flash.rs (48 LoC): flash_alpha(), ease_out(), ease_in(), FLASH_DURATION
- store.rs (1,003 LoC): NotificationStore + impl + tests

lib.rs becomes a 13-line re-export hub. All downstream crates continue
to `use amux_notify::*` unchanged.

Modeled on wezterm's utility crate pattern (separate types, core logic,
and helpers). No behavior change — pure mechanical refactor.
